Refactor pyenv-based auto-install into finder
This should only change one small behaviour: I modified the
post-installation error message to remove the PATH reference because
we're no longer just looking in PATH; the Python user just installed is
simply not located.

The installation itself and when it's triggered is totally unchanged.

NOTE: We probably need to rename the PIPENV_DONT_USE_PYENV flag if we
add other installation methods in the future.